fix: height type was uint32 instead of uint64

+++ clients/platform/v0/objective-c/Platform.pbobjc.h
@@ -78,9 +78,9 @@ typedef GPB_ENUM(ResponseMetadata_FieldNumber) {
 
 @interface ResponseMetadata : GPBMessage
 
-@property(nonatomic, readwrite) uint32_t height;
+@property(nonatomic, readwrite) uint64_t height;
 
-@property(nonatomic, readwrite) uint32_t coreChainLockedHeight;
+@property(nonatomic, readwrite) uint64_t coreChainLockedHeight;
 
 @end
